Format JSON using prettier
This enforces consistent formatting and as a bonus allows us to see
diffs visually on GitHub. Before formatting any change would cause a
diff of "everything replaced", but this new formatted file will show up
as individual changes when diffs are generated.

Formatting is as straightforward as running prettier -w on this file, or
automatically with a plugin such as VS Code's prettier + format
(manually or on save).